Move BaseOperator to `airflow/sdk/bases/operator.py`
For consistency with other modules in `airflow/sdk/bases/`
