+++ airflow-core/src/airflow/decorators/base.py
@@ -41,7 +41,7 @@ from airflow.models.expandinput import (
     ListOfDictsExpandInput,
     is_mappable,
 )
-from airflow.sdk.bases.baseoperator import BaseOperator as TaskSDKBaseOperator
+from airflow.sdk.bases.operator import BaseOperator as TaskSDKBaseOperator
 from airflow.sdk.definitions._internal.contextmanager import DagContext, TaskGroupContext
 from airflow.sdk.definitions.asset import Asset
 from airflow.sdk.definitions.mappedoperator import MappedOperator, ensure_xcomarg_return_value
