+++ task-sdk/src/airflow/sdk/definitions/xcom_arg.py
@@ -34,7 +34,7 @@ from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.xcom import XCOM_RETURN_KEY
 
 if TYPE_CHECKING:
-    from airflow.sdk.bases.baseoperator import BaseOperator
+    from airflow.sdk.bases.operator import BaseOperator
     from airflow.sdk.definitions.edges import EdgeModifier
     from airflow.sdk.types import Operator
